updating file smbios

+++ BootloaderCorePkg/Library/SmbiosInitLib/SmbiosTemplate.c
@@ -247,7 +247,7 @@ SMBIOS_TABLE_TYPE17  mMemoryDevice  = {
   SMBIOS_STRING_INDEX_7,                        // MemoryTechnology;               // The enumeration value from MEMORY_DEVICE_TECHNOLOGY
   {
     sizeof (MEMORY_DEVICE_OPERATING_MODE_CAPABILITY),
-    (UINT8)(BIT1)
+    (UINT16)(BIT1)
   },                                            // MemoryOperatingModeCapability;
   0,                                            // FirwareVersion;
   0,                                            // ModuleManufacturerID;
